复习两数之和，新增合并数组

+++ src/main/java/com/zh/leetcode/array/medium/MedianOfArray.java
@@ -37,7 +37,8 @@ public class MedianOfArray {
 
     }
 
-    public Integer deal(int[] nums1,int[] nums2) {
+    public double deal(int[] nums1,int[] nums2) {
+
 
         return 0;
     }
